+++ frontend/src/pages/participatedMentoring/ParticipatedMentoring.tsx
@@ -1,61 +1,34 @@
+import { useEffect, useState } from 'react';
+
 import styled from '@emotion/styled';
 
+import { getParticipatedMentoringList } from './apis/getParticipatedMentoring';
 import MentoringItem from './MentoringItem/MentoringItem';
 import MentoringList from './MentoringList/MentoringList';
 
 import type { ParticipatedMentoringType } from './types/participatedMentoring';
 
-export const PARTICIPATED_MENTORING_LIST: ParticipatedMentoringType[] = [
-  {
-    mentorName: '이수업',
-    mentorProfileImage:
-      'https://www.google.com/url?sa=i&url=https%3A%2F%2Fwww.crowdpic.net%2Fphoto%2F%25ED%2592%258D%25EA%25B2%25BD-%25EC%259E%2590%25EC%2597%25B0-%25EB%2593%25A4%25ED%258C%2590-%25EC%25B4%2588%25EC%259B%2590-%25EB%2582%2598%25EB%25AC%25B4-136857%3Fsrsltid%3DAfmBOopK3sXndcC9IZQXkHBG_NVOu_ZELeaPahtqFY0gMkGhB9zkxqBr&psig=AOvVaw2i3WaVGdyiSKf3VAcV-BK2&ust=1754548076306000&source=images&cd=vfe&opi=89978449&ved=0CBUQjRxqFwoTCODkj4rH9Y4DFQAAAAAdAAAAABAE',
-    price: '5000',
-    reservedAt: '2024-01-09',
-    categories: ['근력 증진', '다이어트', '요가'],
-    isReviewed: true,
-    status: '완료됨',
-  },
-  {
-    mentorName: '김트레이너',
-    mentorProfileImage:
-      'https://www.google.com/url?sa=i&url=https%3A%2F%2Fwww.crowdpic.net%2Fphoto%2F%25ED%2592%258D%25EA%25B2%25BD-%25EC%259E%2590%25EC%2597%25B0-%25EB%2593%25A4%25ED%258C%2590-%25EC%25B4%2588%25EC%259B%2590-%25EB%2582%2598%25EB%25AC%25B4-136857%3Fsrsltid%3DAfmBOopK3sXndcC9IZQXkHBG_NVOu_ZELeaPahtqFY0gMkGhB9zkxqBr&psig=AOvVaw2i3WaVGdyiSKf3VAcV-BK2&ust=1754548076306000&source=images&cd=vfe&opi=89978449&ved=0CBUQjRxqFwoTCODkj4rH9Y4DFQAAAAAdAAAAABAE',
-    price: '5000',
-    reservedAt: '2024-01-10',
-    categories: ['근력 증진'],
-    isReviewed: false,
-    status: '완료됨',
-  },
-  {
-    mentorName: '박코치',
-    mentorProfileImage:
-      'https://www.google.com/url?sa=i&url=https%3A%2F%2Fwww.crowdpic.net%2Fphoto%2F%25ED%2592%258D%25EA%25B2%25BD-%25EC%259E%2590%25EC%2597%25B0-%25EB%2593%25A4%25ED%258C%2590-%25EC%25B4%2588%25EC%259B%2590-%25EB%2582%2598%25EB%25AC%25B4-136857%3Fsrsltid%3DAfmBOopK3sXndcC9IZQXkHBG_NVOu_ZELeaPahtqFY0gMkGhB9zkxqBr&psig=AOvVaw2i3WaVGdyiSKf3VAcV-BK2&ust=1754548076306000&source=images&cd=vfe&opi=89978449&ved=0CBUQjRxqFwoTCODkj4rH9Y4DFQAAAAAdAAAAABAE',
-    price: '4000',
-    reservedAt: '2024-01-12',
-    categories: ['요가', '필라테스'],
-    isReviewed: false,
-    status: '승인됨',
-  },
-  {
-    mentorName: '이헬퍼',
-    mentorProfileImage:
-      'https://www.google.com/url?sa=i&url=https%3A%2F%2Fwww.crowdpic.net%2Fphoto%2F%25ED%2592%258D%25EA%25B2%25BD-%25EC%259E%2590%25EC%2597%25B0-%25EB%2593%25A4%25ED%258C%2590-%25EC%25B4%2588%25EC%259B%2590-%25EB%2582%2598%25EB%25AC%25B4-136857%3Fsrsltid%3DAfmBOopK3sXndcC9IZQXkHBG_NVOu_ZELeaPahtqFY0gMkGhB9zkxqBr&psig=AOvVaw2i3WaVGdyiSKf3VAcV-BK2&ust=1754548076306000&source=images&cd=vfe&opi=89978449&ved=0CBUQjRxqFwoTCODkj4rH9Y4DFQAAAAAdAAAAABAE',
-    price: '3000',
-    reservedAt: '2024-01-15',
-    categories: ['재활 운동', '스트레칭'],
-    isReviewed: false,
-    status: '승인대기',
-  },
-];
-
 function ParticipatedMentoring() {
+  const [participatedMentoringList, setParticipatedMentoringList] = useState<
+    ParticipatedMentoringType[]
+  >([]);
+
+  useEffect(() => {
+    const fetchParticipatedMentoringList = async () => {
+      const data = await getParticipatedMentoringList();
+      setParticipatedMentoringList(data);
+    };
+
+    fetchParticipatedMentoringList();
+  }, []);
+
   return (
     <StyledContainer>
       <StyledTitle>참여한 멘토링</StyledTitle>
       <StyledWrapper>
         <StyledInfoWrapper>
           <StyledSubTitle>
-            참여한 멘토링 목록 ({PARTICIPATED_MENTORING_LIST.length}건)
+            참여한 멘토링 목록 ({participatedMentoringList.length}건)
           </StyledSubTitle>
           <StyledDescription>
             내가 신청한 멘토링 목록을 확인하고 완료된 멘토링에 대해 리뷰를
@@ -63,11 +36,15 @@ function ParticipatedMentoring() {
           </StyledDescription>
         </StyledInfoWrapper>
         <StyledLine />
-        <MentoringList>
-          {PARTICIPATED_MENTORING_LIST.map((item) => (
-            <MentoringItem key={item.mentorName} mentoring={item} />
-          ))}
-        </MentoringList>
+        {participatedMentoringList.length > 0 ? (
+          <MentoringList>
+            {participatedMentoringList.map((item) => (
+              <MentoringItem key={item.reservationId} mentoring={item} />
+            ))}
+          </MentoringList>
+        ) : (
+          <StyledDescription>참여한 멘토링이 없습니다.</StyledDescription>
+        )}
       </StyledWrapper>
     </StyledContainer>
   );
